refactor(services): turns tabs into routes
Turns the tabs of the Service detail view into nav tabs based on actual routes.
